+++ docs/content/components/content/formatters/percent-format.demo.tsx
@@ -1,0 +1,9 @@
+import { NumericFormat } from '@marigold/system';
+
+export default () => (
+  <NumericFormat
+    styleFormat="percent"
+    value={0.7333}
+    minimumFractionDigits={2}
+  />
+);